+++ src/components/ui/button.tsx
@@ -46,7 +46,13 @@ const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
       <Comp
         className={cn(buttonVariants({ variant, size, className }))}
         ref={ref}
+        type={props.type || "button"}
         {...props}
+        style={{
+          WebkitTapHighlightColor: "transparent",
+          touchAction: "manipulation",
+          ...props.style
+        }}
       />
     );
   }